fix(dev): Add missing Spring AI Testcontainers dependency to MongoDB Atlas module
