+++ arconia-dev/arconia-dev-services/arconia-dev-services-mongodb-atlas/src/main/java/io/arconia/dev/services/mongodb/atlas/package-info.java
@@ -1,4 +1,4 @@
 @NullMarked
-package io.arconia.dev.services.mongodb;
+package io.arconia.dev.services.mongodb.atlas;
 
 import org.jspecify.annotations.NullMarked;
