+++ arconia-dev/arconia-dev-services/arconia-dev-services-mongodb-atlas/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,1 +1,1 @@
-io.arconia.dev.services.mongodb.MongoDbAtlasDevServicesAutoConfiguration
+io.arconia.dev.services.mongodb.atlas.MongoDbAtlasDevServicesAutoConfiguration
