+++ arconia-dev/arconia-dev-services/arconia-dev-services-mongodb/src/test/java/io/arconia/dev/services/mongodb/atlas/MongoDbDevServicesAutoConfigurationTests.java
@@ -1,4 +1,4 @@
-package io.arconia.dev.services.mongodb;
+package io.arconia.dev.services.mongodb.atlas;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.autoconfigure.AutoConfigurations;
