+++ arconia-dev/arconia-dev-services/arconia-dev-services-mongodb/src/test/java/io/arconia/dev/services/mongodb/atlas/MongoDbDevServicesPropertiesTests.java
@@ -1,4 +1,4 @@
-package io.arconia.dev.services.mongodb;
+package io.arconia.dev.services.mongodb.atlas;
 
 import java.time.Duration;
 import java.util.Map;
@@ -10,16 +10,16 @@ import io.arconia.dev.services.core.config.DevServicesProperties;
 import static org.assertj.core.api.Assertions.assertThat;
 
 /**
- * Unit tests for {@link MongoDbAtlasDevServicesProperties}.
+ * Unit tests for {@link MongoDbDevServicesProperties}.
  */
-class MongoDbAtlasDevServicesPropertiesTests {
+class MongoDbDevServicesPropertiesTests {
 
     @Test
     void shouldCreateInstanceWithDefaultValues() {
-        MongoDbAtlasDevServicesProperties properties = new MongoDbAtlasDevServicesProperties();
+        MongoDbDevServicesProperties properties = new MongoDbDevServicesProperties();
 
         assertThat(properties.isEnabled()).isTrue();
-        assertThat(properties.getImageName()).contains("mongodb/mongodb-atlas-local");
+        assertThat(properties.getImageName()).contains("mongo");
         assertThat(properties.getEnvironment()).isEmpty();
         assertThat(properties.getShared()).isEqualTo(DevServicesProperties.Shared.NEVER);
         assertThat(properties.getStartupTimeout()).isEqualTo(Duration.ofMinutes(2));
@@ -27,16 +27,16 @@ class MongoDbAtlasDevServicesPropertiesTests {
 
     @Test
     void shouldUpdateValues() {
-        MongoDbAtlasDevServicesProperties properties = new MongoDbAtlasDevServicesProperties();
+        MongoDbDevServicesProperties properties = new MongoDbDevServicesProperties();
 
         properties.setEnabled(false);
-        properties.setImageName("mongodb/mongodb-atlas-local");
+        properties.setImageName("mongo:latest");
         properties.setEnvironment(Map.of("KEY", "value"));
         properties.setShared(DevServicesProperties.Shared.ALWAYS);
         properties.setStartupTimeout(Duration.ofMinutes(5));
 
         assertThat(properties.isEnabled()).isFalse();
-        assertThat(properties.getImageName()).isEqualTo("mongodb/mongodb-atlas-local");
+        assertThat(properties.getImageName()).isEqualTo("mongo:latest");
         assertThat(properties.getEnvironment()).containsEntry("KEY", "value");
         assertThat(properties.getShared()).isEqualTo(DevServicesProperties.Shared.ALWAYS);
         assertThat(properties.getStartupTimeout()).isEqualTo(Duration.ofMinutes(5));
